+++ README.md
@@ -1,25 +1,20 @@
 ### [구현할 기능 목록들] 5주차 2단계 - 주문하기 
-### 1. 과제 진행 요구사항 수행 
-- [x] Web 플랫폼 등록 > http://localhost:8080 저장
-
-### 2. 카카오톡 메시지 API를 사용하여 주문하기 기능 구현
-- [x] 주문할 때 수령인에게 보낼 메시지를 작성
+### 1. 카카오톡 메시지 API를 사용하여 주문하기 기능 구현
+(1) 카카오 로그인을 통해 인가코드 받고 토큰까지 받기. 그리고 DB에 저장하기
+- [x] application-secret.yml 파일에 카카오 로그인 정보를 저장한다.
+- [x] 카카오 로그인을 통해 인가코드를 받고 sendRedirect를 통해 토큰값을 받는다.
+- [x] access token값을 통해 사용자 이름을 가져온다.
+- [x] 사용자 이름과 토큰값을 DB에 저장한다.
+(2) 주문 요청 및 주문 내역 저장
 - [x] 상품 옵션과 해당 수량을 선택하여 주문하면 해당 상품 옵션의 수량이 차감된다.
+  - [x] 단, 옵션 id에 해당하는 옵션이 있는지 확인한다.
+  - [x] 옵션의 수량이 충분한지 확인한다.
 - [x] 해당 상품이 위시 리스트에 있는 경우 위시 리스트에서 삭제한다.
-- [ ] 나에게 보내기를 읽고 주문 내역을 카카오톡 메시지로 전송한다. 
+- [x] 주문 내역인 옵션 id, 수량, 메시지를, 주문 시간을 저장한 후 반환한다.
+(3) 주문 내역을 카카오톡 메시지로 나에게 전송하기
+- [x] 주문 내역을 카카오톡 메시지로 전송한다. 
 
-### 3. refactor 및 고도화 
-(1) 카카오 로그인을 통해 액세스 토큰 받고 토큰값과 사용자 이름 저장하기 
-- [x] RestClient를 활용하여 인가 코드를 받은 후 redirect해주기
-- [x] 받은 인가 코드를 활용하여 액세스 토큰을 받기
-- [x] 토큰을 통해 사용자 이름를 가져오기
-- [x] 토큰과 사용자 이름 DB에 저장하기 
-(2) 주문하고 주문서를 나에게 보내기
-- [ ] 옵션 아이디, 수량, 메시지를 request body로 받기
-- [ ] 상품을 판매할 수 있을지 판단하고 삭제하기
-- [ ] 카카오 API 서버에서 메시지를 보내기 
-
-
+### 2. refactor 및 고도화
 - [ ] 옵션 이름 한국어 안되는 오류 해결 
 - [ ] 옵션 정보 페이지 수정/삭제 버튼 css 수정 
-
+- [ ] accesstoken값을 통해 주문 내역에 저장된 사용자 id를 확인하여 유효한지 확인하기
