+++ src/main/java/gift/order/service/KakaoService.java
@@ -1,10 +1,14 @@
 package gift.order.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import gift.common.config.KakaoProperties;
+import gift.option.domain.OrderResponse;
 import gift.order.domain.Token;
 import gift.order.dto.KakaoUser;
 import gift.order.dto.KakaoTokenResponse;
+import gift.order.dto.Link;
+import gift.order.dto.TemplateObject;
 import gift.order.repository.TokenJPARepository;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.http.HttpHeaders;
@@ -13,6 +17,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClient;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -22,7 +27,7 @@ import java.net.URI;
 @Service
 public class KakaoService {
     private final KakaoProperties kakaoProperties;
-    private final RestClient client = RestClient.builder().build();
+    private final RestClient restClient = RestClient.builder().build();
     private final ObjectMapper objectMapper = new ObjectMapper();
     private final TokenJPARepository tokenJPARepository;
     public KakaoService(KakaoProperties kakaoProperties, TokenJPARepository tokenJPARepository) {
@@ -38,8 +43,8 @@ public class KakaoService {
         // body
         LinkedMultiValueMap<String, String> body = createBody(code);
 
-        // response
-        KakaoTokenResponse kakaoTokenResponse = client.post()
+        // RestClient 사용하여 POST 요청
+        KakaoTokenResponse kakaoTokenResponse = restClient.post()
                 .uri(URI.create(url))
                 .contentType(MediaType.APPLICATION_FORM_URLENCODED)
                 .body(body)
@@ -66,8 +71,8 @@ public class KakaoService {
     public KakaoUser getUserInfo(String accessToken) throws IOException {
         String url = "https://kapi.kakao.com/v2/user/me";
 
-        // response
-        String response = client.get()
+        // RestClient 사용하여 GET 요청
+        String response = restClient.get()
                 .uri(URI.create(url))
                 .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
                 .retrieve()
@@ -92,5 +97,31 @@ public class KakaoService {
         Token token = new Token(accessToken, refreshToken, userName);
 
         tokenJPARepository.save(token);
+    }
+
+    public void sendKakaoMessage(OrderResponse orderResponse, String accessToken) throws JsonProcessingException {
+        // validate
+        if (accessToken == null || accessToken.isEmpty()) {
+            throw new IllegalArgumentException("[ERROR] 유효하지 않은 액세스 토큰입니다.");
+        }
+
+        // 카카오톡 메시지 보내기
+        ObjectMapper objectMapper = new ObjectMapper();
+        TemplateObject templateObject = new TemplateObject(
+                "text",
+                orderResponse.toString(),
+                new Link("link")
+        );
+        String templateObjectJson = objectMapper.writeValueAsString(templateObject);
+        MultiValueMap<Object, Object> map = new LinkedMultiValueMap<>();
+        map.set("template_object", templateObjectJson);
+
+        restClient.post()
+                .uri(kakaoProperties.getMessageToMeUri())
+                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
+                .header("Authorization", "Bearer " + accessToken)
+                .body(map)
+                .retrieve()
+                .body(String.class);
     }
 }
